+++ web/app/[locale]/(home)/_ui/lock/component.tsx
@@ -3,6 +3,8 @@
 import { useState } from 'react';
 import styles from './styles.module.css';
 
+import type { Seed } from '@/actions/types';
+
 const MIN = 1;
 const MAX = 99;
 
@@ -14,19 +16,27 @@ function randomArray(length: number, min: number, max: number) {
   return Array.from({ length }, () => randomInt(min, max));
 }
 
-export function PlayButton({
-  reset,
-  onPlay,
+export function LockDoor({
+  vortex,
+  disabled = false,
+  reset = true,
+  onAttempt,
 }: {
-  onPlay?: (seed: number) => void;
+  vortex?: boolean;
+  disabled?: boolean;
+  onAttempt?: (match: boolean, seed: Seed) => void;
   reset?: boolean;
 }) {
+  const [holding, setHolding] = useState<boolean>(true);
+
   const [result, setResult] = useState<number>(NaN);
   const [pending, setPending] = useState<boolean>(false);
   const [values, setValues] = useState<number[]>(randomArray(8, MIN, MAX));
+  const [match, setMatch] = useState(false);
+
   const [intervalRef, setIntervalRef] = useState<NodeJS.Timeout | null>(null);
   const [playTimeout, setPlayTimeout] = useState<number>(1000);
-  const [match, setMatch] = useState(false);
+  const [restartRef, setRestartRef] = useState<NodeJS.Timeout | null>(null);
 
   const shiftAndPush = () => {
     setValues((values) => randomArray(values.length, MIN, MAX));
@@ -34,13 +44,21 @@ export function PlayButton({
   };
 
   const start = () => {
+    if (pending || match || disabled) return;
+
+    setHolding(false);
+    restartRef && clearTimeout(restartRef);
+
     setResult(NaN);
     shiftAndPush();
 
-    setIntervalRef(setInterval(shiftAndPush, 100));
+    setIntervalRef(setInterval(shiftAndPush, 200));
   };
 
   const play = () => {
+    if (pending || match || disabled) return;
+
+    const timestamp = Date.now();
     setPending(true);
     intervalRef && clearInterval(intervalRef);
 
@@ -50,44 +68,44 @@ export function PlayButton({
       const match = value % values.length === pos - 1;
       setResult(pos);
       setMatch(match);
-      onPlay && onPlay(value);
       setPending(false);
+      onAttempt && onAttempt(match, { value, trigger: playTimeout, timestamp });
 
       reset &&
-        match &&
-        setTimeout(() => {
-          setMatch(false);
-          setResult(NaN);
-        }, 5000);
+        !match &&
+        setRestartRef(
+          setTimeout(() => {
+            setMatch(false);
+            setResult(NaN);
+            setHolding(true);
+          }, 5000)
+        );
     }, playTimeout);
   };
 
   return (
     <button
-      className={`${styles.border} bg-[url('/img/push-me/play.svg')] bg-center bg-no-repeat bg-cover`}
+      className={`${styles.border} bg-[url('/assets/images/entry/lock/full.svg')] bg-center bg-no-repeat bg-cover`}
       onMouseDown={start}
       onMouseUp={play}
       onTouchStart={start}
       onTouchEnd={play}
-      disabled={pending || match}
+      disabled={pending || match || disabled}
       data-pending={pending}
       data-match={match}
+      data-vortex={vortex}
       data-revealed={!Number.isNaN(result)}
       data-result={result || undefined}
     >
       <div className={`${styles.door} w-full h-full`}>
-        <div className={`${styles.button}`}>
-          Push
-          <br />
-          Me
-        </div>
         {values.map((value, i) => (
           <div
             key={i}
             className={[
               styles.value,
+              holding ? styles.holding : '',
               result === i + 1 ? styles.active : '',
-              value % values.length === i ? styles.match : '',
+              !holding && value % values.length === i ? styles.match : '',
             ].join(' ')}
             style={{ gridArea: `i${i + 1}` }}
             data-area={`i${i + 1}`}