+++ web/components/shared/index.ts
@@ -1,3 +1,3 @@
 export * from './CreateTokenAccount';
-export * from './PlayButton';
+export * from '@/app/[locale]/(home)/_ui/lock';
 export * from './Price';