+++ web/providers/index.ts
@@ -1,3 +1,5 @@
+export * from './bg-provider';
+
 export * from './react-query-provider';
 export * from './crypto-provider';
 export * from './lucky-bags-provider';
@@ -5,6 +7,6 @@ export * from './wallet-provider';
 export * from './solana-provider';
 
 export * from '@/components/cluster/cluster-data-access';
-export * from '@/app/games/_provider/games-provider';
+export * from '@/app/console/games/_provider/games-provider';
 
 export * from './data-feed';