+++ web/tailwind.config.js
@@ -6,12 +6,15 @@ module.exports = {
   content: [
     join(
       __dirname,
-      '{src,pages,components,app}/**/*!(*.stories|*.spec).{ts,tsx,html}'
+      '{src,pages,components,app,ui}/**/*!(*.stories|*.spec).{ts,tsx,html}'
     ),
     ...createGlobPatternsForDependencies(__dirname),
   ],
   theme: {
     extend: {
+      screens: {
+        // xs: '428px',
+      },
       boxShadow: {
         // glow: '0 0 10px #0f0, 0 0 20px #0f0, 0 0 30px #0f0, 0 0 40px #0f0',
         glow: '0 0 20px 10px #00f, 0 0 30px 15px #00f, 0 0 40px 20px #00f, 0 0 50px 25px #00f',