+++ web/ui/icons/index.ts
@@ -1,0 +1,7 @@
+export * from './IconSet';
+
+export * from './Social';
+export * from './DexScreener';
+export * from './Jupiter';
+
+export * from './Money';